Create get_sensor_data.py
For now, im able to get the values from the commands, do proper parsing but assembling the curl for the API is still a todo. Trash code. Main point is to get this to work as it should. 

Had problems with the dumb return code of the MPL13115A2 sensor. Even with the correct values, it throws an exit code of 50. Filed an issue at https://github.com/ControlEverythingCommunity/MPL3115A2/issues/2

This is more comments than code. FML.